+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 > [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-> [Semantic Versutilning](https://semver.org/spec/v2.0.0.html).
+> [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0]
+### Added
+ - CallStack.getCallerOf() overload so that it operates on `StackTraceElement`
 
 ## [0.1.0]
 ### Added
@@ -12,5 +16,6 @@
  - LICENSE
  - README.md
 
-[Unreleased]: https://github.com/hankadler/kotlin-util-callstack/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/hankadler/kotlin-util-callstack/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/hankadler/kotlin-util-callstack/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/hankadler/kotlin-util-callstack/releases/tag/v0.1.0